@@ -23,25 +23,28 @@
   "license": "ISC",
   "dependencies": {
     "@cycle/run": "^3.1.0",
+    "disqus-react": "^1.0.9",
     "express": "^4.14.0",
     "grunt": "^1.2.1",
     "himalaya": "^0.2.0",
+    "immutable": "^4.0.0-rc.12",
     "moment": "^2.15.0",
     "postcss-calc": "^5.3.1",
     "postcss-custom-media": "^5.0.1",
-    "react": "^15.3.1",
-    "react-datetime": "^2.6.0",
-    "react-dimensions": "^1.3.0",
-    "react-disqus-thread": "^0.4.0",
-    "react-document-meta": "^2.0.3",
-    "react-dom": "^15.3.1",
+    "prop-types": "^15.7.2",
+    "react": "^16.13.1",
+    "react-datetime": "^2.16.3",
+    "react-dimensions": "^1.3.1",
+    "react-document-meta": "^3.0.0-beta.2",
+    "react-dom": "^16.13.1",
     "react-images": "^0.5.1",
-    "react-immutable-proptypes": "^2.1.0",
-    "react-redux": "^4.4.5",
-    "react-router": "^2.8.1",
-    "react-screen-orientation": "0.0.2",
+    "react-immutable-proptypes": "^2.2.0",
+    "react-player": "^2.6.0",
+    "react-redux": "^7.2.1",
+    "react-router": "^5.2.0",
+    "react-router-dom": "^5.2.0",
+    "react-screen-orientation": "^0.0.4",
     "react-share": "^1.9.3",
-    "react-youtube": "^7.2.0",
     "redux": "^3.6.0",
     "redux-cycles": "^0.4.0",
     "redux-immutable": "^3.0.8",
