@@ -13,77 +13,76 @@
  * The App container. It is responsible for rendering the matching route
  */
 // Do this once before any other code in your app (http://redux.js.org/docs/advanced/AsyncActions.html)
-import 'babel-polyfill'
-import React, { Component, PropTypes } from 'react'
-import { render } from 'react-dom';
-import * as actions from '../actions/document'
+import 'babel-polyfill';
+import PropTypes from 'prop-types';
+import React, {Component} from 'react';
+import * as actions from '../actions/document';
 import {connect} from 'react-redux';
-import DeviceOrientation, { Orientation } from 'react-screen-orientation'
-import {isBrowser} from '../utils/appHelpers'
-import ImmutablePropTypes from 'react-immutable-proptypes'
-import globalStyles from '../assets/styles/global.css';
+import ImmutablePropTypes from 'react-immutable-proptypes';
+import {useLocation} from 'react-router-dom';
 
 // The children are the components of the chosen route
 class App extends Component {
 
-    componentDidMount() {
-        this.props.setDocumentLocation(this.props.location)
-    }
-    componentWillReceiveProps(nextProps){
-        // Set the current document
-        if (this.props.documentKey != nextProps.documentKey)
-            this.props.showDocument(nextProps.documentKey)
+  componentDidMount() {
+    this.props.setDocumentLocation(this.props.location);
+  }
 
-        // If the location hash came in the URL set that
-        if (this.props.location.hash != nextProps.location.hash) {
-            this.props.setDocumentLocation(nextProps.location)
-        }
-    }
+  componentWillReceiveProps(nextProps) {
+    // Set the current document
+    if (this.props.documentKey != nextProps.documentKey)
+      this.props.showDocument(nextProps.documentKey);
 
-    render() {
-        return React.cloneElement(this.props.children, {})
+    // If the location hash came in the URL set that
+    if (this.props.location.hash != nextProps.location.hash) {
+      this.props.setDocumentLocation(nextProps.location);
     }
+  }
+
+  render() {
+    return React.cloneElement(this.props.children, {});
+  }
 };
 
 App.PropTypes = {
-    documentKey: PropTypes.string,
-    overlayDocumentKey: PropTypes.string,
-    document: ImmutablePropTypes.map,
-    location: React.PropTypes.object,
-    orientation: React.PropTypes.bool
-}
+  documentKey: PropTypes.string,
+  overlayDocumentKey: PropTypes.string,
+  document: ImmutablePropTypes.map,
+  location: PropTypes.object,
+  orientation: PropTypes.bool
+};
 
 function mapStateToProps(state, props) {
 
-    // The key of the document if one was specified in the URL
-    const splat = props.params.splat
-    const documents = state.getIn(['documents', 'entries'])
-    const overlayDocumentKey = state.getIn(['documents', 'currentOverlay'])
+  // The key of the document if one was specified in the URL
 
-    let document = null
-    if (splat && documents.get(splat)) {
-        document = documents.get(splat)
-    }
-    else {
-        // Load and show the newest post that isn't marked in the future
-        // TODO this will need to be used by a post listing
-        const sortedDocuments = state.getIn(['documents', 'entries']).sort(
-            (a, b) => a.get('date') > b.get('date'))
-        const now = new Date();
-        // Get the newest document. Ignore documents like About and Contact that have no date
-        document = sortedDocuments.reverse().find(document => document.get('date') && document.get('date') <= now)
-    }
-    // Add the window.orientation so we can detect orientation in iOS devices
-    return Object.assign(state.toObject(), {
-        documentKey: documents.keyOf(document),
-        document: document,
-        overlayDocumentKey,
-        orientation: window.orientation
-    })
+  const splat = props.location.splat;
+  const documents = state.getIn(['documents', 'entries']);
+  const overlayDocumentKey = state.getIn(['documents', 'currentOverlay']);
+
+  let document = null;
+  if (splat && documents.get(splat)) {
+    document = documents.get(splat);
+  } else {
+    // Load and show the newest post that isn't marked in the future
+    // TODO this will need to be used by a post listing
+    const sortedDocuments = state.getIn(['documents', 'entries']).sort(
+      (a, b) => a.get('date') > b.get('date'));
+    const now = new Date();
+    // Get the newest document. Ignore documents like About and Contact that have no date
+    document = sortedDocuments.reverse().find(document => document.get('date') && document.get('date') <= now);
+  }
+  // Add the window.orientation so we can detect orientation in iOS devices
+  return Object.assign(state.toObject(), {
+    documentKey: documents.keyOf(document),
+    document: document,
+    overlayDocumentKey,
+    orientation: window.orientation
+  });
 
 }
 
 export default connect(
-    mapStateToProps,
-    actions
-)(App)
+  mapStateToProps,
+  actions
+)(App);
